requirements.txt
pyTelegramBotAPI
openai
requests
yt-dlp
python-dotenv